Updates entities and repositories
Applies JsonIgnoreProperties to entities to handle lazy initialization issues during serialization.

Changes the GenerationType for idProducto in Producto and idJuegoDeMesa in JuegoDeMesa to IDENTITY.

Removes the OneToOne relationship between JuegoDeMesa and Producto, replacing it with a Long field idProducto in JuegoDeMesa.

Updates AdminRepository to find by both correo and password.

Adds methods to ProductoRepository for finding products by location, name, and stock. Adds a query to retrieve products with the most stock.

Updates RankingRepository to use int instead of Long for ID.

Updates service implementations to adapt to entity changes.

+++ Grupo_11/src/main/java/com/DBD/Grupo_11/Repository/RankingRepository.java
@@ -6,5 +6,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface RankingRepository extends JpaRepository<Ranking,Integer> {
-    Ranking findById(Long id);
+    Ranking findById(int id);
 }
